# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CurrentOrganizationResponse:
    properties:
      alias:
        type: string
      id:
        type: string
      roles:
        items:
          $ref: '#/definitions/middleware.OrganizationRole'
        type: array
    type: object
  handler.CurrentUserResponse:
    properties:
      globalRoles:
        items:
          $ref: '#/definitions/middleware.GlobalRole'
        type: array
      organization:
        $ref: '#/definitions/handler.CurrentOrganizationResponse'
      organizations:
        items:
          $ref: '#/definitions/handler.CurrentOrganizationResponse'
        type: array
      subject:
        type: string
      username:
        type: string
    type: object
  middleware.GlobalRole:
    enum:
    - admin
    - moderator
    - visitor
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleModerator
    - RoleVisitor
  middleware.OrganizationRole:
    enum:
    - org_admin
    - event_manager
    - finance_viewer
    type: string
    x-enum-varnames:
    - RoleOrganizationAdmin
    - RoleEventManager
    - RoleFinanceViewer
  model.APIError:
    properties:
      error:
        $ref: '#/definitions/model.APIErrorDetail'
    type: object
  model.APIErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  model.ConfigureOrgRolesRequest:
    properties:
      roles:
        items:
          $ref: '#/definitions/model.OrganizationRole'
        type: array
        uniqueItems: true
    required:
    - roles
    type: object
  model.ConfigureOrgRolesResponse:
    properties:
      message:
        type: string
      organizationId:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  model.CreateOrganizationRequest:
    properties:
      alias:
        type: string
      city:
        type: string
      contactEmail:
        type: string
      contactPhoneNumber:
        type: string
      countryCode:
        type: string
      houseNumber:
        type: string
      name:
        type: string
      orgAdmin:
        type: string
      postalCode:
        type: string
      street:
        type: string
    required:
    - alias
    - name
    - orgAdmin
    type: object
  model.CreateOrganizationResponse:
    properties:
      alias:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      detail:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  model.OrganizationRole:
    enum:
    - org_admin
    - event_manager
    - finance_viewer
    type: string
    x-enum-varnames:
    - RoleOrganizationAdmin
    - RoleEventManager
    - RoleFinanceViewer
...
info:
  contact: {}
  description: REST API for the EventHub platform
  title: EventHub API
  version: "1.0"
paths:
...
  /organizations/:
    post:
      consumes:
      - application/json
      description: Creates a new organization in Keycloak and the API database. Requires
        the admin role on the backend client.
      parameters:
      - description: Organization to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CreateOrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - organizations
  /organizations/{organizationID}/members/{username}/roles:
    put:
      consumes:
      - application/json
      description: 'Replaces the complete role set of an organization member. Requires
        the org_admin role in the given organization (global admins bypass this check);
        it is checked against the token and again against Keycloak, so a demoted admin
        cannot act on an old token. organizationID is the Keycloak organization ID
        or alias as returned by POST /organizations and GET /users/me. roles is required;
        an empty array removes all roles. Roles: org_admin (manage members), event_manager
        (manage events), finance_viewer (view sales and billing). Removing the last
        org_admin is rejected. The affected user must refresh their token before the
        new roles take effect.'
      parameters:
      - description: Keycloak organization ID or Alias
        in: path
        name: organizationID
        required: true
        type: string
      - description: Username of the organization member
        in: path
        name: username
        required: true
        type: string
      - description: Complete role set for the member (empty array = no roles)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ConfigureOrgRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ConfigureOrgRolesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Configure organization member roles
      tags:
      - organizations
  /users/me:
    get:
      description: Returns the authenticated user's profile, global roles, active
        organization, and organization memberships
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CurrentUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" where {token} is a Keycloak access token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"